+++ c/src/lib/libbsp/i386/pc386/ne2000/ne2000.c
@@ -1073,7 +1073,7 @@ ne_stats (struct ne_softc *sc)
 /* NE2000 driver ioctl handler.  */
 
 static int
-ne_ioctl (struct ifnet *ifp, int command, caddr_t data)
+ne_ioctl (struct ifnet *ifp, u_long command, caddr_t data)
 {
   struct ne_softc *sc = ifp->if_softc;
   int error = 0;
